One gradle-plugins module only
