+++ gradle-plugins/build.gradle.kts
@@ -5,25 +5,25 @@ plugins {
 gradlePlugin {
     plugins.register("kmp-platform-plugin") {
         id = "kmp-platform-plugin"
-        implementationClass = "com.gchristov.thecodinglove.kmpgradleplugins.KmpPlatformPlugin"
+        implementationClass = "com.gchristov.thecodinglove.gradleplugins.kmp.KmpPlatformPlugin"
     }
     plugins.register("kmp-module-plugin") {
         id = "kmp-module-plugin"
-        implementationClass = "com.gchristov.thecodinglove.kmpgradleplugins.KmpModulePlugin"
+        implementationClass = "com.gchristov.thecodinglove.gradleplugins.kmp.KmpModulePlugin"
+    }
+    plugins.register("build-config-plugin") {
+        id = "build-config-plugin"
+        implementationClass = "com.gchristov.thecodinglove.gradleplugins.BuildConfigPlugin"
     }
     plugins.register("javascript-browser-executable-plugin") {
         id = "javascript-browser-executable-plugin"
         implementationClass =
-            "com.gchristov.thecodinglove.kmpgradleplugins.JavascriptBrowserExecutablePlugin"
+            "com.gchristov.thecodinglove.gradleplugins.js.JavascriptBrowserExecutablePlugin"
     }
     plugins.register("javascript-node-executable-plugin") {
         id = "javascript-node-executable-plugin"
         implementationClass =
-            "com.gchristov.thecodinglove.kmpgradleplugins.JavascriptNodeExecutablePlugin"
-    }
-    plugins.register("build-config-plugin") {
-        id = "build-config-plugin"
-        implementationClass = "com.gchristov.thecodinglove.kmpgradleplugins.BuildConfigPlugin"
+            "com.gchristov.thecodinglove.gradleplugins.js.JavascriptNodeExecutablePlugin"
     }
 }
