--- kmp/kmp-gradle-plugins/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "kmp-gradle-plugins"
